Use autoflush and fix LogPrinter during testing
