+++ tests/BaseTestCase.php
@@ -3,6 +3,8 @@ declare(strict_types=1);
 
 namespace Tests;
 
+use Paraunit\Configuration\EnvVariables;
+use Paraunit\File\Cleaner;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -11,6 +13,9 @@ use PHPUnit\Framework\TestCase;
  */
 class BaseTestCase extends TestCase
 {
+    /** @var string|null */
+    private $randomTempDir;
+
     protected function getCoverageStubFilePath(): string
     {
         $filename = __DIR__ . '/Stub/CoverageOutput/CoverageStub.php';
@@ -35,5 +40,35 @@ class BaseTestCase extends TestCase
     protected function getStubPath(): string
     {
         return realpath(__DIR__ . DIRECTORY_SEPARATOR . 'Stub') . DIRECTORY_SEPARATOR;
+    }
+
+    protected function createRandomTmpDir()
+    {
+        $this->randomTempDir = uniqid(sys_get_temp_dir() . DIRECTORY_SEPARATOR . 'paraunit-test-', true);
+        $this->randomTempDir .= DIRECTORY_SEPARATOR;
+
+        $this->assertTrue(
+            putenv(EnvVariables::LOG_DIR . '=' . $this->randomTempDir),
+            'Failed setting env variable for log dir'
+        );
+    }
+
+    protected function getRandomTempDir(): string
+    {
+        $this->assertNotNull($this->randomTempDir, 'Tmp dir not initialized');
+        
+        return $this->randomTempDir;
+    }
+
+    protected function tearDown()
+    {
+        putenv(EnvVariables::LOG_DIR);
+        putenv(EnvVariables::PROCESS_UNIQUE_ID);
+
+        if ($this->randomTempDir && is_dir($this->randomTempDir)) {
+            Cleaner::cleanUpDir($this->randomTempDir);
+        }
+
+        parent::tearDown();
     }
 }
