+++ tests/Functional/Parser/JSON/LogPrinterTest.php
@@ -3,8 +3,7 @@ declare(strict_types=1);
 
 namespace Tests\Functional\Parser\JSON;
 
-use Paraunit\Configuration\StaticOutputPath;
-use Paraunit\Configuration\TempFilenameFactory;
+use Paraunit\Configuration\EnvVariables;
 use Paraunit\Parser\JSON\LogPrinter;
 use PHPUnit\Framework\TestSuite;
 use Tests\BaseFunctionalTestCase;
@@ -15,25 +14,11 @@ use Tests\BaseFunctionalTestCase;
  */
 class LogPrinterTest extends BaseFunctionalTestCase
 {
-    public function testLogFilenameMatches()
+    protected function setUp()
     {
-        $testName = __CLASS__;
-        $testSuite = $this->prophesize(TestSuite::class);
-        $testSuite->getName()
-            ->willReturn($testName);
-        $testSuite->count()
-            ->willReturn(1);
+        parent::setUp();
 
-        $dir = sys_get_temp_dir();
-        new StaticOutputPath($dir);
-        $printer = new LogPrinter();
-
-        $reflectionMethod = new \ReflectionMethod($printer, 'getLogFilename');
-        $reflectionMethod->setAccessible(true);
-        $this->assertEquals(
-            $dir . DIRECTORY_SEPARATOR . md5(__FILE__) . '.json.log',
-            $reflectionMethod->invoke($printer, $testSuite->reveal())
-        );
+        $this->createRandomTmpDir();
     }
 
     public function testWrite()
@@ -44,26 +29,19 @@ class LogPrinterTest extends BaseFunctionalTestCase
             ->willReturn($testName);
         $testSuite->count()
             ->willReturn(1);
-        $logFilename = $this->getLogFilenameForTest(__FILE__);
 
-        new StaticOutputPath(dirname($logFilename));
+        putenv(EnvVariables::PROCESS_UNIQUE_ID . '=' . md5(__FILE__));
+        $logFullPath = $this->getRandomTempDir() . md5(__FILE__) . '.json.log';
+
         $printer = new LogPrinter();
 
         $printer->startTestSuite($testSuite->reveal());
 
-        $this->assertFileExists($logFilename);
+        $this->assertFileExists($logFullPath);
 
-        $content = file_get_contents($logFilename);
+        $content = file_get_contents($logFullPath);
         $this->assertJson($content);
         $decodedJson = json_decode($content, true);
         $this->assertEquals(['event' => 'suiteStart', 'suite' => $testName, 'tests' => 1], $decodedJson);
-    }
-
-    private function getLogFilenameForTest(string $testFilename): string
-    {
-        /** @var TempFilenameFactory $filenameFactory */
-        $filenameFactory = $this->container->get('paraunit.configuration.temp_filename_factory');
-
-        return $filenameFactory->getFilenameForLog(md5($testFilename));
     }
 }
