+++ tests/Unit/Parser/JSON/LogPrinterTest.php
@@ -18,9 +18,6 @@ use Tests\BaseUnitTestCase;
  */
 class LogPrinterTest extends BaseUnitTestCase
 {
-    /** @var string */
-    private $randomTempDir;
-
     public function testStartTestSuite()
     {
         $this->createPrinterAndStartTestSuite();
@@ -292,7 +289,7 @@ class LogPrinterTest extends BaseUnitTestCase
 
     private function getLogContent(): string
     {
-        $logFilename = $this->randomTempDir . 'log-file-name.json.log';
+        $logFilename = $this->getRandomTempDir() . 'log-file-name.json.log';
         $this->assertFileExists($logFilename, 'Log file missing! Maybe you called this method too early?');
 
         $content = file_get_contents($logFilename);
@@ -322,25 +319,5 @@ class LogPrinterTest extends BaseUnitTestCase
             'suite' => get_class($this),
             'tests' => 1,
         ];
-    }
-
-    private function createRandomTmpDir()
-    {
-        $this->randomTempDir = uniqid(sys_get_temp_dir() . DIRECTORY_SEPARATOR . 'paraunit-test-', true);
-        $this->randomTempDir .= DIRECTORY_SEPARATOR;
-
-        putenv(EnvVariables::LOG_DIR . '=' . $this->randomTempDir);
-    }
-
-    protected function tearDown()
-    {
-        putenv(EnvVariables::LOG_DIR);
-        putenv(EnvVariables::PROCESS_UNIQUE_ID);
-
-        if (is_dir($this->randomTempDir)) {
-            Cleaner::cleanUpDir($this->randomTempDir);
-        }
-
-        parent::tearDown();
     }
 }
